[mmf] update producer module according to previous PR
* fix kvs producer busy loop when audio is enabled

Issue: When audio is enabled, the putMedia loop won't send frame unless there are available video and audio frames. It makes the loop isn't empty and doesn't sleep.

Solution: Instead of check if the stream buffer is empty, it can be fixed by checking the send count. If the send count is zero, then it means there is no action in current loop and it can sleep a little.

>> contributed by williamlai (Amazon)